predictor/requirements.txt
fastapi
uvicorn
pandas
numpy
joblib
kafka-python
scikit-learn
